Refactor other methods out of createSystem
