+++ packages-next/keystone/src/lib/createGraphQLSchema.ts
@@ -1,0 +1,70 @@
+import { GraphQLSchema, GraphQLObjectType } from 'graphql';
+import { mergeSchemas } from '@graphql-tools/merge';
+import { mapSchema } from '@graphql-tools/utils';
+import type { KeystoneConfig, SessionStrategy } from '@keystone-next/types';
+import { adminMetaSchemaExtension } from '@keystone-next/admin-ui/templates';
+
+import { gql } from '../schema';
+
+export function createGraphQLSchema(
+  config: KeystoneConfig,
+  keystone: any,
+  adminMeta: any,
+  sessionStrategy?: SessionStrategy<unknown>,
+  sessionImplementation?: any
+) {
+  // @ts-ignore
+  const server = keystone.createApolloServer({
+    schemaName: 'public',
+    dev: process.env.NODE_ENV === 'development',
+  });
+  const schemaFromApolloServer: GraphQLSchema = server.schema;
+  const schema = mapSchema(schemaFromApolloServer, {
+    'MapperKind.OBJECT_TYPE'(type) {
+      if (
+        config.lists[type.name] !== undefined &&
+        config.lists[type.name].fields._label_ === undefined
+      ) {
+        let {
+          fields: { _label_, ...fields },
+          ...objectTypeConfig
+        } = type.toConfig();
+        return new GraphQLObjectType({ fields, ...objectTypeConfig });
+      }
+
+      return type;
+    },
+  });
+
+  // TODO: find a way to not pass keystone in here, if we can - it's too broad and makes
+  // everything in the keystone instance public API
+  let graphQLSchema = config.extendGraphqlSchema?.(schema, keystone) || schema;
+  if (sessionStrategy?.end) {
+    graphQLSchema = mergeSchemas({
+      schemas: [graphQLSchema],
+      typeDefs: gql`
+        type Mutation {
+          endSession: Boolean!
+        }
+      `,
+      resolvers: {
+        Mutation: {
+          async endSession(rootVal, args, context) {
+            await context.endSession();
+            return true;
+          },
+        },
+      },
+    });
+  }
+  graphQLSchema = adminMetaSchemaExtension({
+    adminMeta,
+    graphQLSchema,
+    isAccessAllowed:
+      sessionImplementation === undefined
+        ? undefined
+        : config.ui?.isAccessAllowed ?? (({ session }) => session !== undefined),
+    config,
+  });
+  return graphQLSchema;
+}
